Deeplasia-master/requirements.txt

flask==2.2.5
flask-cors==3.0.10
PyYAML==6.0.1
reportlab==3.6.12


torch==1.10.2
torchvision==0.11.3
pytorch-lightning==1.6.4
torchmetrics==0.7.0
albumentations==1.1.0


opencv-python==4.6.0.66
pillow==9.5.0
numpy==1.22.0
pandas==1.3.5
scikit-learn==1.0.2
scikit-image==0.19.1
scipy==1.7.3
seaborn==0.11.2
matplotlib==3.5.3
